Add watson service to handle connection, response and parsing for Watson Personality Insights API call
