+++ spec/fixtures/cassettes/user-logout.yml
@@ -29,20 +29,20 @@ http_interactions:
       Content-Type:
       - application/json;charset=utf-8
       Date:
-      - Fri, 01 Jun 2018 03:02:44 GMT
+      - Sat, 02 Jun 2018 01:14:36 GMT
       Expires:
       - Tue, 31 Mar 1981 05:00:00 GMT
       Last-Modified:
-      - Fri, 01 Jun 2018 03:02:44 GMT
+      - Sat, 02 Jun 2018 01:14:36 GMT
       Pragma:
       - no-cache
       Server:
       - tsa_a
       Set-Cookie:
-      - guest_id=v1%3A152782216487208582; Expires=Sun, 31 May 2020 03:02:44 GMT; Path=/;
+      - guest_id=v1%3A152790207696902242; Expires=Mon, 01 Jun 2020 01:14:36 GMT; Path=/;
         Domain=.twitter.com
-      - personalization_id="v1_nFnD77EQ2+HUFBfvzRyERA=="; Expires=Sun, 31 May 2020
-        03:02:44 GMT; Path=/; Domain=.twitter.com
+      - personalization_id="v1_OttB7mjg5LTudpfzvQ+xtg=="; Expires=Mon, 01 Jun 2020
+        01:14:36 GMT; Path=/; Domain=.twitter.com
       Status:
       - 200 OK
       Strict-Transport-Security:
@@ -50,7 +50,7 @@ http_interactions:
       X-Access-Level:
       - read
       X-Connection-Hash:
-      - 3cabc5bbe9e2e7821beb20c87e53391f
+      - 3462ae27ccb55feca3ffac8b2e1a0389
       X-Content-Type-Options:
       - nosniff
       X-Frame-Options:
@@ -58,13 +58,13 @@ http_interactions:
       X-Rate-Limit-Limit:
       - '900'
       X-Rate-Limit-Remaining:
-      - '892'
+      - '894'
       X-Rate-Limit-Reset:
-      - '1527822409'
+      - '1527902947'
       X-Response-Time:
-      - '23'
+      - '21'
       X-Transaction:
-      - 005e44b700b7d612
+      - '009bc51b00947e54'
       X-Twitter-Response-Tags:
       - BouncerCompliant
       X-Xss-Protection:
@@ -74,5 +74,5 @@ http_interactions:
       string: '{"id":809075773105721344,"id_str":"809075773105721344","name":"\u219fA\u0274\u1d05\u0280e\u1d21","screen_name":"agpiermarini","location":"","profile_location":null,"description":"\ud83c\udf43","url":null,"entities":{"description":{"urls":[]}},"protected":false,"followers_count":0,"friends_count":0,"listed_count":0,"created_at":"Wed
         Dec 14 16:41:11 +0000 2016","favourites_count":0,"utc_offset":null,"time_zone":null,"geo_enabled":false,"verified":false,"statuses_count":0,"lang":"en","contributors_enabled":false,"is_translator":false,"is_translation_enabled":false,"profile_background_color":"000000","profile_background_image_url":"http:\/\/abs.twimg.com\/images\/themes\/theme1\/bg.png","profile_background_image_url_https":"https:\/\/abs.twimg.com\/images\/themes\/theme1\/bg.png","profile_background_tile":false,"profile_image_url":"http:\/\/pbs.twimg.com\/profile_images\/968045861958635520\/LxDeesSo_normal.jpg","profile_image_url_https":"https:\/\/pbs.twimg.com\/profile_images\/968045861958635520\/LxDeesSo_normal.jpg","profile_link_color":"ABB8C2","profile_sidebar_border_color":"000000","profile_sidebar_fill_color":"000000","profile_text_color":"000000","profile_use_background_image":false,"has_extended_profile":false,"default_profile":false,"default_profile_image":false,"following":null,"follow_request_sent":null,"notifications":null,"translator_type":"none"}'
     http_version: 
-  recorded_at: Fri, 01 Jun 2018 03:02:44 GMT
+  recorded_at: Sat, 02 Jun 2018 01:14:37 GMT
 recorded_with: VCR 4.0.0
