+++ spec/search/twitter_profile_search_spec.rb
@@ -15,10 +15,10 @@ describe TwitterProfileSearch do
           expect(subject.profile.name).to eq('Turing School')
           expect(subject.profile.location).to eq('Denver, CO')
           expect(subject.profile.description).to eq('We unlock human potential by training a diverse, inclusive student body to succeed in high-fulfillment technical careers.')
-          expect(subject.profile.followers_count).to eq(2749)
+          expect(subject.profile.followers_count).to eq(2750)
           expect(subject.profile.following_count).to eq(587)
-          expect(subject.profile.last_tweet).to eq("RT @a_reuster: 2 more reasons why @turingschool is awesome: 1. We make sure our students have solid foundations before we move them on 2. O…")
-          expect(subject.profile.tweet_count).to eq(3894)
+          expect(subject.profile.last_tweet).to eq("RT @ameseee: I'm fortunate to work at a place where I'm pushed technically every day and still feel connected to K12 education! Inspired by…")
+          expect(subject.profile.tweet_count).to eq(3895)
           expect(subject.profile.avatar_url).to eq("http://pbs.twimg.com/profile_images/444146105312243712/EHdy1GyV_400x400.png")
         end
       end
