+++ spec/services/watson_service_spec.rb
@@ -1,0 +1,8 @@
+require 'rails_helper'
+
+describe WatsonService do
+  subject { WatsonService.new("data") }
+  it 'initializes with data' do
+    expect(subject).to be_a WatsonService
+  end
+end
